Create psutil/_constants.py
