+++ MANIFEST.in
@@ -27,6 +27,7 @@ include docs/make.bat
 include docs/requirements.txt
 include psutil/__init__.py
 include psutil/_common.py
+include psutil/_constants.py
 include psutil/_ntuples.py
 include psutil/_psaix.py
 include psutil/_psbsd.py
