+++ psutil/_common.py
@@ -9,7 +9,6 @@ psutil or third-party modules.
 """
 
 import collections
-import enum
 import functools
 import os
 import socket
@@ -43,8 +42,6 @@ __all__ = [
     'CONN_CLOSE', 'CONN_CLOSE_WAIT', 'CONN_CLOSING', 'CONN_ESTABLISHED',
     'CONN_FIN_WAIT1', 'CONN_FIN_WAIT2', 'CONN_LAST_ACK', 'CONN_LISTEN',
     'CONN_NONE', 'CONN_SYN_RECV', 'CONN_SYN_SENT', 'CONN_TIME_WAIT',
-    # net constants
-    'NIC_DUPLEX_FULL', 'NIC_DUPLEX_HALF', 'NIC_DUPLEX_UNKNOWN',  # noqa: F822
     # process status constants
     'STATUS_DEAD', 'STATUS_DISK_SLEEP', 'STATUS_IDLE', 'STATUS_LOCKED',
     'STATUS_RUNNING', 'STATUS_SLEEPING', 'STATUS_STOPPED', 'STATUS_SUSPENDED',
@@ -117,24 +114,6 @@ CONN_LISTEN = "LISTEN"
 CONN_CLOSING = "CLOSING"
 CONN_NONE = "NONE"
 
-
-# net_if_stats()
-class NicDuplex(enum.IntEnum):
-    NIC_DUPLEX_FULL = 2
-    NIC_DUPLEX_HALF = 1
-    NIC_DUPLEX_UNKNOWN = 0
-
-
-globals().update(NicDuplex.__members__)
-
-
-# sensors_battery()
-class BatteryTime(enum.IntEnum):
-    POWER_TIME_UNKNOWN = -1
-    POWER_TIME_UNLIMITED = -2
-
-
-globals().update(BatteryTime.__members__)
 
 # --- others
