+++ psutil/_psbsd.py
@@ -27,6 +27,8 @@ from ._common import conn_to_ntuple
 from ._common import debug
 from ._common import memoize
 from ._common import memoize_when_activated
+from ._constants import BatteryTime
+from ._constants import NicDuplex
 
 __extra__all__ = []
 
@@ -295,8 +297,7 @@ def net_if_stats():
             if err.errno != errno.ENODEV:
                 raise
         else:
-            if hasattr(_common, 'NicDuplex'):
-                duplex = _common.NicDuplex(duplex)
+            duplex = NicDuplex(duplex)
             output_flags = ','.join(flags)
             isup = 'running' in flags
             ret[name] = ntp.snicstats(isup, duplex, speed, mtu, output_flags)
@@ -339,9 +340,9 @@ if FREEBSD:
             return None
         power_plugged = power_plugged == 1
         if power_plugged:
-            secsleft = _common.POWER_TIME_UNLIMITED
+            secsleft = BatteryTime.POWER_TIME_UNLIMITED
         elif minsleft == -1:
-            secsleft = _common.POWER_TIME_UNKNOWN
+            secsleft = BatteryTime.POWER_TIME_UNKNOWN
         else:
             secsleft = minsleft * 60
         return ntp.sbattery(percent, secsleft, power_plugged)
